requirements.txt
alembic==1.13.2
fastapi==0.115.4
httpx==0.27.2
Pillow==10.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35
uvicorn[standard]==0.30.6
